Rewrite column ID tests to focus on behavior, not infrastructure
Remove tests that inspect InMemoryTable internals (ID assignment,
uniqueness, complex type IDs) without using a DataFrame. Remove tests
unrelated to column IDs (TABLE_ID_MISMATCH, COLUMNS_MISMATCH).

Keep and group behavioral tests: mismatch detection, complex types,
joins, temp views, writes, null table/column ID connectors.

Add missing test for multiple column ID mismatches.

Move design doc scenarios and externalInsert helper to
DataSourceV2ExternalSessionSuite alongside the extSession tests.
